service::deploy::gitclone: change clone path to match repo name
Currently if a repo is under a path other than the root, the clone
will succeed but there will be a mismatch between the repo path and
the actual cloned repo - for example, /srv/deployment/aqs is the
cloned version of the analytics/aqs repo. This is not a problem in and
of itself, but it becomes a problem when the init script for aqs uses
$repo as the path to find the repo itself. Rather than modify the
heavily used service::node module, adapt gitclone to be in line with
the non-git deployments.

Change-Id: Ia022a00c5fe1aba42c7fb6954b213cd8c2c3780e