ceph: add ceph repo and parameter to all client modules
This adds the specific ceph repository to all the machines that
include ceph::common (or any of the clients, ex. cloudvirts).

This includes adding the parameter to all the clients, and
refactoring the apt repo to be in the ceph::common class (and some
tests).

Bug: T274566
Hosts: re:^cloud.*(wmnet|org)$
Change-Id: I6c3e08f0c4bbff4ceef968a644aa29199e1e6121
Signed-off-by: David Caro <dcaro@wikimedia.org>