kube-apiserver: Update admission controller config
The -–admission-control flag was deprecated in kubernetes 1.10 and replaced
with -–enable-admission-plugins. As kube-apiserver enables a set of
default admission plugins, we no longer need to specify the complete
list in the class or hiera.

The list of controllers enabled by default can be optained from:
"kube-apiserver -h | grep enable-admission-plugins"

To disable one of the default admission controllers, you may use
$disable_admission_plugins now.

This will enable the following additional admission controllers (moving
closer to kubernetes defaults):
- TaintNodesByCondition
- MutatingAdmissionWebhook
- ValidatingAdmissionWebhook
- RuntimeClass

This also modifies the config templates to use multiline DAEMON_ARGS and
sorted lists as in arguments where applicable to make spotting
differences easier.

Bug: T270063
Change-Id: I56d5486a70ecc559f7d1ed07ebedfbacc8825a2e