policy.yaml files: update default behavior
It seems that the 'default' policy rule is not always observed
in the newly-refactored policy modules.

Our default rules exist to replace a default behavior of
'all members can do it' with 'only projectadmins can do it'.

We can accomplish that by changing the admin_or_owner rule
to actually check for projectadmin rather than simply
project membership.

Neutron actually observes the default rule, so need to change
anything for Neutron.

Change-Id: Iede65d44c9a4abb433007ad910f44b0ff515d238