cache: enable exp caching policy on upload@eqsin
For the last few days we have been using the 'exp' cache admission
policy on cp5001 with rate=0.1 and base=-20.3. With such settings, the
probability of objects being cached based on their size is as follows:

  16 KB 98.6%
 128 KB 89.4%
1024 KB 40.8%
4096 KB 2.78%
5120 KB 1.13%
8192 KB 0.07%

Extend the policy change to the rest of upload@eqsin. Set
large_objects_cutoff to 8M so that objects larger than 4M have a chance
to be admitted, albeit small.

Bug: T275809
Hosts: cp5001.eqsin.wmnet,cp5002.eqsin.wmnet
Change-Id: I9da97cd434f0ad9f63604c4ce58927773f09e624