P:debmonitor::server: Drop nginx in favour of tlsproxy
First attempt of this change resulted in the default apache page, this
happened because the cas site only listens on debmonitor.wikimedia.org.
With the nginx configuration we use the following which means apache
sees the external domain
 proxy_set_header Host $host;
however envoy will use the host header present in the request it sees
which in this case is the `internal_server_name` as such add this as a
server aliases.

Change-Id: Ic61e1c87e057cb580d4f3370b701322777924289
Hosts: P:debmonitor::server