aptrepo: add component libmysql-java to buster-wikimedia
The hive server and metastore daemons (analytics) do not work with the
libmariadb-java jdbc driver, since the sql used in the prepared
statements seems not compatible with what mariadb provides. While we
follow up with Apache Hive upstream, we need to unblock the Buster
upgrade of the Analytics cluster, so this change adds a component to
buster-wikimedia to forward port libmysql-java to it.

Bug: T278424
Change-Id: I9ad7376c0348aeef8f84662864847f681aafe26d