bigtop::mysql_jdbc: use component/libmysql-java for buster
We have created a new component for buster-wikimedia called
'component/libmysql-java', and added the last version of the package
from Debian Stretch. This is needed to allow the Hive Metastore to avoid
failures when issuing sql prepared statements to mariadb, since for some
reason the mariadb-specific jdbc driver is not compatible (it leads to
syntax errors while executing the statements).

Bug: T278424
Change-Id: Id87ba81e8c7dfc32c03fcd9ccd412007ddb2ca99