sonofgridengine: grid-configurator: introduce support for the new domain
Configure new hosts only with the new domain.

The new domain is mandatory for all new VMs in CloudVPS. This script runs on
sgemaster servers, which are already Buster / new domain VMs.

For compatibility with older VMs running Stretch / old domain, some special
case is taken into account:

* before this patch, if there was a VM using both the old and new domain
(i.e, an already configured Stretch VM, which is reported by Nova as using the
new domain) then the script would have tried to delete the old FQDN and add
the new one. This results in an add/remove dancing that eventually fails,
because there are queues (and jobs) on that node.

* with this patch, we detect this case, and ignore a VM with both old and new
domain, to prevent the artificial back-and-forth, the eventual script failure
and hopefully any other disruption.

* this still allows the script to detect missing nodes (old or new domain) and
to remove them from the grid as usual.

This patch also updates the tests to match the new behavior.

Bug: T277653
Change-Id: Id0884ce09ab31109b788d757291e4bd9f9e39241
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>