Setup new etcd3 cluster on conf200[456] in codfw
The configcluster_stretch role is buster now, so remove that from the
description at first. After migrating, we can get rid of the _stretch
suffix in the role name as well.

Exclude the zookeeper roles on those nodes to allow for a two step
migration. Zookeeper may be added back once etcd is done.

Bug: T271573
Change-Id: I4cdf8d1e545c19e8c3c442b979475199237c9877