Move sanitize_eventlogging_analytics jobs from data_purge to refine_santizie
This will allows to more easily declare RefineSanitize jobs as needed.

Aside from some puppet and systemd timer name changes, this should be
a functional no-op.  A future patch will refactor the sanitize jobs
to match what is currently in test/refine_sanitize.pp

This change does rename the salt used for sanitization, so that
the to be declared jobs can all use the same salt.
The existent eventlogging sanitization salt will be copied manually
to the new name.

Bug: T273789
Change-Id: I966d0a59a08dc63a4b561601b56304da49469cac