cache: enable exp caching policy on upload@ulsfo
We recently enabled the exp caching policy in the cache_upload Singapore
cluster (22ac208) with very promising results. Cache usage improved
significantly, resulting in an increase in frontend hitrate of about 3%.

To ensure that larger objects have a chance to get cached properly once
the cache is full if they make the probabilisitc threshold, we also
increased nuke_limit with b1e2345.

Apply the policy in San Francisco as well to evaluate the impact there.

Bug: T275809
Hosts: cp4021.ulsfo.wmnet
Change-Id: Ifad9d66196e24b38f9a48c726bad00e4cb432993