hadoop: improve default log4j config
This change is a continuation of what done for the hdfs audit log. It
extends the same rolling gz appender to the security log file and the
hadoop file. The former is used by HDFS Namenodes and Journalnodes,
meanwhile the latter is used by all hadoop daemons. The result will be a
drastic reduction of the size used by logs, not only on the Master
nodes.

This change also adds more buffer for the hdfs audit log (only used by
the namenodes).

Bug: T276906
Change-Id: Ibde8082cc39921437c6c4fce7c1b2a0dc10f565e