sudo: add new flag purge_sudoeres_d
I recently noticed that if a user removes a sudo::user from the puppet
manifest, without first marking it absent, then the old sudo
rule will be left in place no the system.  This PS adds a new parameter
to the sudo class which allows puppet to manage the ./etc/sudoers.d
folder and purge any files which are not directly manage by puppet

This CR also moves the sudo class out of the admin class and into
profile::standard.

Hosts: cumin1001.eqiad.wmnet
Change-Id: Id0741cc151a350a3a43aae5a4b456ff59e020ddf