Only disable timers for ipmitoo/smartmon timers up to Buster
Up to 0.17 prometheus-node-exporter shipped a number of collectors in
text_collector_examples/, which were also shipped in the Debian package.
These were eventually moved to a separate repository (
https://github.com/prometheus-community/node-exporter-textfile-collector-scripts)
and got packaged as src:prometheus-node-exporter-collectors. Starting with
0.18.1+ds-2 (and thus Bullseye) prometheus-node-exporter declares a Recommends:
on that package. But we don't install recommended packages by default and we don't
need any of the shipped collectors, so restrict the disabling of the broken timers
to distros up to Buster

Bug: T275873
Change-Id: Icbd47ab14a81c03aaefbafa1ee6f98bf4ecd2d99