[Beta Cluster] mediawiki: Remove "commons.wikipedia" redirect
This aligns the beta config more closely with production.

In 2014, this redirect was copied from prod with commit 816fc32fb842,
but it was never specifically needed in Beta. The production version has
since been removed in favour of `redirects.dat`. That service isn't
provisioned yet in Beta afaik, but either way we don't need this.

Change-Id: I94578c21122065c0986590dfdf10933f99d72bed