lists: Port check_mailman_queue to Python
The main motivation for this is that I want to add mailman3 support (in
a follow-up). It now mentions which queue is too large:
 CRITICAL: 1 mailman2 queues above limits: virgin is 28 (limit: 25)

It's overall more defensive too (e.g. limits are validated to be
integers). Any exception will be caught and output as an error instead
of a generic unknown failure.

Change-Id: I4eeb17d7862284b098e92d7406833d7dba767314