dbbackups: Add enabled parameter backups on the cluster::management role
Because we can sometimes have more than one host with the cluster::
management role per datacenter at a time, have a way to disable its
scheduling of new backups.

A new parameter "enabled" has been added to the dbbackups::transfer
profile, that defaults to true, but allows now disabling of the
active systemd timers.

While nothing horribly broke if there was multiple scheduling backup
host per datacenter (either the missing configuration would make
puppet complain, or one could add an empty configuration and the
backups would run as a noop), it is nice to have a way to enable
and disable the functionality on specific hosts/roles/datacenters.

Bug: T138562
Change-Id: I4991b2ed0b93f6ee39a75d10b79bf9f167b2f73c