Move WCQS update to Tuesday
Wikimedia Commons dump, which we use to reload WCQS is finishing
recently after we start the weekly update, which means we get the
old one, which is almost a week old at this point. This patch
moves update to Tuesday - temporary measure until we get a
continuous process in place (T262265).

Bug: T280022
Change-Id: I697b6c5c5face2b7c07a03481c0b489f86b31294