sanitize_eventlogging_analytics_immediate - ensure absent during switchover
We have started renaming tables in event_sanitized, and the
sanitize_eventlogging_analytics_immediate EventLoggingSanitization job
will no longer do the right thing with output paths.

We need to switch to RefineSanitze and a newer refinery version,
but fisrt we need to ensure that all tables in event_sanitized are
properly renamed.

Bug: T280813
Bug: T273789
Change-Id: I249a968e18b261d9d6d1362aa3a698e1cf60b3b6