cloudstore: set up secondary_drbd classes
In order to move the secondary cluster to a more sustainable DRBD model,
refactor the puppet so that it is possible to apply a new role/profile
during the setup phase of DRBD. Once the first server is configured
for DRBD properly, the next step is to enable the syncserver tool,
copying the data over via rsync. When that is done the standalone DRBD
host becomes the primary (failover), and then the new role can be applied
to the old primary to convert that system to DRBD as well.

The syncserver setup will be dropped once replication over DRBD is confirmed.

It is hoped that this role will handle a lot of the config, but manual
steps are likely.

Bug: T224747
Change-Id: I7f3a0ed28dcbc03465052514270a0c52d7ae627c