profile::thanos::swift: add account for ML serve cluster
The Kubeflow cluster that Machine Learning is building will need a Swift
account to fetch models to serve. The Swift cluster chosen to support
the MVP project is Thanos, since it support the S3 API and it is
replicated in both eqiad and codfw. Medium to long term we'll surely
need to migrate to a more appropriate solution like MOSS.

The private part of this commit will be added once the naming is
accepted, just to avoid extra commits.

Bug: T280773
Change-Id: I3656f75823ddec71fade6451c88f1cede75116ce