toolforge: Add all tools created before 2020-07-07 to legacy_redirector
The redirector that we use to remap URLs for the legacy
https://tools.wmflabs.org/{tool} naming scheme was initial configured to
support only the tools that were actively running either a grid engine
or kubernetes web service on the date that the config was generated.
This has led to "gaps" in the data caused by tools with legacy URLs
published which were for one reason or another not actively running
a webservice on the given date.

This patch replaces the list of allowed tools with a hopefully
definitive list based on LDAP data for tools. The list was created using
this shell invocation:

  $ ldapsearch \
    -xLLL -P 3 -E pr=40000/noprompt -o ldif-wrap=no -b"dc=wikimedia,dc=org" \
    '(&(cn=tools.*)(objectClass=posixGroup)(createTimestamp<=20200707000000Z))' \
    cn |
    grep 'cn:' |
    awk '{sub(/^[^.]*\./, "", $2); print $2}' |
    sort |
    sed "s/^/'/;s/$/',/"

This LDAP search finds all posixGroup objects with names that start with
"tools." which were created on or before 2020-07-07T00:00Z.

Bug: T281003
Change-Id: Ie60ab8e0e97295e8930852f4c45899e30fafb99c