systemd::timer::job: update the program name of systemd-timer-mail-wrapper
When systemd::timer jobs need to send out emails we use wrap the command
using the `systemd-timer-mail-wrapper` script.  When we do this it means
that the program name that hits the jornal is
`systemd-timer-mail-wrapper` and not `demonitor-client` which prevents
the rsyslogd rule from making the correct match.

This CR updates the script so that it first renames it self to the name
of the command its wrapping

Hosts: cumin1001.eqiad.wmnet
Bug: T281090
Change-Id: I4de460094a179f79e2dc0d33df7a3616a59027a3