dbbackups: Add s2 to db1102, a buster backup source
In order to prepare the eventual switchover of s2 to 10.4, setup
a backup source host with buster with s2. db1102, which already
contained an instance of x1, now also will have an s2 instance.

This host will be provisioned with the data from db1156, which
has me restored through a logical dump.

Also disable temporarily notifications while data is transferred.

Bug: T280492
Bug: T280979
Change-Id: Iaa4051d261a28183d699d0e2d2c1dc2f7b47c6f5