openstack.nova: libvirt use systemd socket instead of --listen
It turns out that on systemd-enabled oses it's the current default and
recommended way.

```
In  socket  activation mode, it will rely on systemd to create and
listen on the UNIX, and optionally TCP/IP, sockets and pass them as
pre-opened file descriptors. In this mode, it is not permitted to
pass the --listen parameter, and most of the socket related config
options in /etc/libvirt/libvirtd.conf will no longer have any effect.
To enable TCP or TLS sockets use either

          $ systemctl start libvirtd-tls.socket

       Or

          $ systemctl start libvirtd-tcp.socket
```

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Iabbb07fe54d1de8066c15112ebc086b49597e39b