cache: Safari regex refinement
It turns out that Google Chrome has "Safari" in the UA string. For
example:

  [...] Chrome/90.0.4430.85 Safari/537.36

Actual Safari 14 on OSX has the following in User-Agent:

  [...] Version/14.0.3 Safari/605.1.15

While on iOS the string looks like this:

  [...] Version/14.0.3 Mobile/15E148 Safari/604.1

Given that Chrome has no issues with WebP files to the best of our
knowledge, it seems reasonable to make the WebP regex a bit stricter by
using "Safari/6" instead.

Bug: T280439
Change-Id: Iba75653917b25f02b8ce7eaea217db7786ad5610