test/data_purge - add drop_event job
This job drops all data in the Hive event database
older than 90 days.

We should apply this job in the main cluster ONLY
AFTER we have properly copied and are sanitizing all
table that should be kept in the event_sanitized database.

This patch includes the data_purge jobs on an-test-coord1001,
which apparently previously have never been present!

Bug: T273789
Change-Id: I7591b0de383c0c533b295b0fee9bddb9af9f46ea