ci::master/deployment_server: add new k8s namespace for shellbox
Like I58aab6fca485 this is following:

https://wikitech.wikimedia.org/wiki/Pipeline_admin#Adding_a_new_namespace_and_service

with the example change:

https://github.com/wikimedia/puppet/commit/dd0616c05f928a55a586034e4e6b760298504712

but minus the default-kubernetes-policy.yaml file which is not in this
repo anymore.

It will also need a change in the private repo to go with it but besides
that it should be the very first step to get shellbox into kubernetes.

Bug: T260330
Change-Id: I0f48049dab055a2a0ec66c34e694add1c213112c