mailman3: Properly import templates for admins
The import process only imports the templates into mailman3, so they
exist on disk and in config, but are entirely invisible to listadmins.

We really need to import them into postorius, but upstream doesn't have
that functionality yet. So we do it!

* Identify templates on disk
* Directly insert them into mailman3web.postorius_emailtemplate
* Use REST API to point mailman3 at the postorius templates
* Delete templates off disk

For already imported lists, we can do something like:

>>> from migrate_to_mailman3 import fix_templates
>>> fix_templates(listname)

Bug: T281425
Change-Id: I9d94caed930dcae5f9f40fc9baf30757254b70b8