vcl: make vcl_hit invoke default VCL
We have so far unconditionally returned "deliver" in vcl_hit, thereby
skipping the default vcl_hit implementation shipped with the builtin
VCL. The assumption was that "vcl_hit" would be called upon cache hits
only, which depending on the definition of "cache hit", as well as the
version of Varnish, may or may not necessarily always be true. Returning
deliver in case of hit on a stale object results in the object being
sent to the client while a background fetch to refresh it is
automatically initiated by Varnish.

As part of the work done by upstream to address the infamous issue #1799
there have been important changes between Varnish 6.0.0 and 6.0.1, which
include the guarantee that hits on expired objects out of grace now go
straight to vcl_miss and avoid a detour through vcl_hit provided that
there is no other ongoing backend request to refresh that object
already. This means that, starting with Varnish 6.0.1, if the result of
a lookup finds an object in cache which is past its TTL, past grace
time, but still in cache because it is either in keep or for other
corner cases such as the expiry thread lagging behind its mailbox due to
various reasons, the next state after lookup is guaranteed to be miss
instead of hit. In other words, my understanding is that starting with
6.0.1 the distinction between grace and keep is now explicitly
implemented in Varnish core instead of being left to the VCL.

Considering all the above, and additionally the general idea that the
builtin VCL is aware of and takes into account the implementation
details of the Varnish version it has been designed for, make sure we do
invoke the builtin VCL from vcl_hit. With this change applied, there
should be functionally no difference at all on Varnish 6.0.1 and later,
while previous versions (or those with PR 2705 reverted) should now
properly enforce the distinction between grace and keep.

https://web.archive.org/web/20160809123924/https://varnish-cache.org/trac/ticket/1799
https://github.com/varnishcache/varnish-cache/pull/2705/

Bug: T264398
Change-Id: I1c25b2ecd5bb44b3fb5d16dc9c617985f844f2bd