P:debmonitor::client: dont specify an explicit CA
If no CA is present then debmonitor-client uses the default certificate
store.  As the puppet CA is in the certificate store we dont need to
explicitly set it in the debmonitor.conf file.

This change will make exchanging the server SSL to one signed by the
debmonitor intermediate CA (which is also in the /etc/ssl/certs[1]) much
smoother

Change manually tested on cumin1001.eqiad.wment

Hosts: cumin1001.eqaid.wmwnet
Change-Id: I0e16fa3e39e5333fcfae53bf5df9f2c6c12c70ec