P:debmonitor::server: switch debmonitor.discovery.wmnet to cfssl
This change instructs debmonitor to request it certificate from cfssl.

We have already pushed a change to instruct the debmonitor client to
 rely on /etc/ssl/cert instead of specifying a certificate explicitly[1]
We have also installed the root and debmonitor intermediate certificate
to /etc/ssl/certs.  As such this change should deploy with no
interruption, as long as enough time has been left for [1] to deploy
(i.e. don't merge until after 13:20 UTC)

The prep work to enable this was preformed in
  I228a61da1b3910709bec0e1b1c96b412b40d5b72

[1]https://gerrit.wikimedia.org/r/c/operations/puppet/+/683591
[2]https://gerrit.wikimedia.org/r/c/operations/puppet/+/683583

Hosts: P:debmonitor::server
Change-Id: I653751ec36a4e57374ce85568cdbae070219b35a