rsync::quickdatacopy: ensure a destination host gets an rsync client
When rsync::quickdatacopy is used in a role that is applied on both
a source and a destination host, the source host gets rsyncd installed.

(if $source_host == $::fqdn)

But there was nothing that specifically made sure that rsync client
is installed at all in the other case (if $dest_host == $::fqdn).

On a new bullseye host I noticed how on the host that is supposed
to pull FROM the source host, I did not have rsync installed at all.

It makes me think before bullseye it might have been installed by default
or pulled from something else that it now doesn't do anymore.

Also I am doing a completely unrelated "replace require_packages with
ensure_packages" for the stunnel package just because i didn't like to
introduce inconsistency and there is T266479.

Hosts: cumin:R:rsync::quickdatacopy
Bug: T280989
Bug: T266479
Change-Id: Ifbe30b417d62e4c2449a8310ab0eeccb32e34a07