role::analytics_cluster::hadoop::ui: use the 'hue' db
The 'hue_next' database name is confusing, it was used when testing the
new version but now we should use the more canonical 'hue' name (that is
also the one that we backup from analytics-meta).
This change needs to be merged after dropping the current hue database
(belonging to the old instance), dumping hue_next to a new hue, and
finally dropping hue_next.

Bug: T280262
Change-Id: I69f420652bc9ad90cbe346f9d35f702094c71dea