gerrit: escape URI in Phabricator comment
Whenever abandonning a patch with a reason mentioning an URL, it is not
escaped and the URL is decoded when posted. That leads to `+` to be
written back in Phabricator as a space ` `.

So when abandonning a change with the reason:

  https://example.org/+/foo.html

The comment shows up in Phabricator with a broken URL:

  https://example.org/ /foo.html

We have hit the same issue a while back with the subject variable and
passed it to `escapeUri`:
https://groups.google.com/g/closure-templates-discuss/c/lXOqqmkO7fE

Thus escape any URI that might be given in the reason for abandonning a
change.

Bug: T280197
Change-Id: I133f20b33aef7de4d15ca9cd050672ec6fdbe607