wikireplicas: redirect all database CNAMEs to the new system
This should cut over all database CNAMES in regular use to the new
system in such a way that allows the old section names to be retired
entirely from DNS.

Applied the same to the old labsdb names because I'd rather not have
a truly deprecated scheme be a backdoor around the redirect. Additionally
removed the names for the arbcom wikis as they aren't on the wikireplica
servers (private).

Bug: T260389
Change-Id: I5d026798da25e54574bff33753b63d8310885bde