mediawiki: Remove 'deployment.wikimedia' vhost from Beta Cluster
The deletion of this wiki was first proposed in July 2018.
The wiki has been "closed" since June 2020.

As of writing, all references to it that I could find in Codesearch,
Horizon config, mediawiki.org, Meta-Wiki, and Wikitech wiki, have
been removed or updated. (The exception being one example value in
a horizon script, per I7f36ef8f6c).

This patch is the first and most visible step in the actual deletion,
which is the traffic layer from the web. Other decom steps can follow
later, and this is also the easiest to undo.

Bug: T198673
Change-Id: I49016545536756f5d2eb8135f0510e6f14206500