maintain_dbusers: add new multi-instance analytics dedicated host
Since this is all on one host, it could be added via hard coding if this
doesn't work as expected. I'm hoping the heredoc string will work just
like a single-line string. I plan on stopping the service immediately
after merge to be sure the resulting config looks like expected.

Bug: T281287
Change-Id: I9f6bd847e0d764b47af3c54ad242ef639d2c94ec