cloudmetrics: fail over to cloudmetrics1001
To take cloudmetrics1002 out of service, it needs to become the standby.
Also, there has to be a better way than this...

If statsd doesn't know the hostname, the client-side references probably
should go to prometheus-labmon.eqiad.wmnet instead, but I'm not sure.

The server-side parts are about replication and make more sense.

Bug: T281881
Change-Id: I88cc64d5adebbd8e75dbfc2ec1597a2f037ea6bc