bigtop::hadoop::nodemanager: apply systemd override to service
The Nodemanager unit runs in a cgroup with TaskMax (user+kernel threads
counter) set with an heuristic based on kernel version + cpus available.
The Nodemanager's containers allocate a lot of threads, and we have been
seeing issues like the following:
1) MaxTasks reached and Nodemanager failing to allocate native threads.
2) FATAL errors registered by the Nodemanager, and shutdown due to a
   degraded state.
3) Icinga alerts and puppet runs to bring up the service again.

In the Nodemanager case it seems ok to raise the TaskMax contraint,
since we want the service to use as many threads as it needs.
We could use the 'infinity' value to avoid any constraint, but we may
want to be careful too since there are other daemons running on the
worker nodes. The actual MaxTask highest value is 11059 across
workers, so the new value seems to be a good compromise to see if the
current issue can be fixed or not.

Bug: T281792
Change-Id: If37e6ab9b65f0efc0dcb8a638104738244863069