P:pki::client: fix chain an chained certificates
Envoy at the very least needs the chained file to be
 - client cert
 - int cert
instead of the other way round.  further int certs are in a predefined
location in puppet so no need to fetch the via http. finnally we should
not have the root cert in the chain.

Hosts: cumin1001.eqiad.wmnet,debmonitor1002.eqiad.wmnet
Change-Id: I8e9593c42b36d92aeba906363f9aaa67f4498961