P:pki::client: fix chain an chained certificates
Envoy at the very least needs the chained file to be
 - client cert
 - int cert
instead of the other way round.  further int certs are in a predefined
location in puppet so no need to fetch the via http. finnally we should
not have the root cert in the chain.

Change-Id: Ib742b78084af3af3eef0201dbe855074db4337a5
Hosts: cumin1001.eqiad.wmnet,debmonitor1002.eqiad.wmnet