Don't install python-etcd starting with bullseye
python-etcd was introduced in d446e5119515f20dbfff2d52e60fed8abaabb2e0
as a dependency for etcd-manage. This script no longer exists, but let's
keep python-etcd in the manifest for older distros since other code might
implicitly depend on it's presence. If it should still be needed on
bullseye it can be fixed by declaring the package in the respective Puppet
class.

Change-Id: I3926bb4c4baa7baec939bea7f69136a96523423b