wikireplicas: cut over the last IPs to the new cluster
Currently, certain the old eqiad.wmflabs A records for the section names
(eg. s1, s2, s3) are the only things pointing at the old multi-source
cluster. Those should be set to the same IP addresses as the new names
so that all DNS points at the new cluster, period. The old one is not
sticking around.

Bug: T260389
Change-Id: I9818a8a84350664b9c0dbaf73fc8b9a4e3e4d5dc