mariadb: Install 10.5 client on host with only client packages
What's next for RDBMSs at Wikimedia is a decision that is far
in the future. Quoting marostegui:

> which needs a loooot more time

On the client, though, things should be easy to change, as almost
all clients can connect to all kinds of servers.

Enforce 10.5 client cli as the default client package installed
on client-only hosts for the bullseye Debian version (11). This
could be reevaluated later.

Bug: T276589
Change-Id: Idb74a3d5a7cdd3d46c9ef17e1feca76bd35156a1