trafficserver: Fine tune acme-chief cert warning
Reduce from 30 to 21 days the warning threshold for acme-chief issued
certificates. Acme-chief will issue the new certificate 30 days before
the expiration date of the current one and stage it for 7 days to avoid
clock skew issues on the UAs. This staging period is triggering false warnings
for the unified and the wikiworkshop certificates.

Change-Id: I780bb8b656a3eea9ec6dae96d72836b40c7e9f92