dbbackups: Split dbbackups::content into 2 roles, client and storage
dbbackups::content used to host both the generation of dumps and its
long term storage for the other datacenter.

Split the roles so that backup*002 is now only the location where
dumps are generated, and backup*003 where they are stored encrypted
for long term, so we can increase the service capacity (which was
running low).

Also setup the new backup jobs as weekly- to prevent having
redundant full and differential backups, every Wednesday (a day
after dumps run) and increase max size to 80TB per device.

Bug: T282249
Change-Id: I33bbe9028aff2fb93f651aa14956bb6973054dbc