wikireplicas-dns: condense repeated nodes for better failover
Since it is possible that we may want to do a DNS-based failover for
maintenance and downtime on the VM replicas, repeated values should
be a simple repeat annotation in yaml so that you only change one set
of IP addresses in such an operation. That should reduce error and
confusion.

Bug: T260389
Change-Id: I5b79baea81f1e9ad0c80b13bc7833f381fb3b872