profile::hadoop::master: add alert to bump the heap size when needed
We have been following a heap bump increase for the HDFS Namenode
suggested by Cloudera/Hortonworks, that it is based on number of files
stored on HDFS. The idea is to trigger a warning/alert if we cross the
threshold and we don't bump the heap. The main side effects of not doing
it may be heap pressure and slowdowns, so better to be proactive.

Change-Id: Ieb39ebe96a93db6479d4737f14ccc9ae92b4ebe0