cloudgw: don't use concatenation with CIDR
There is a bug in this kernel version, see upstream mailing list for reference:
 https://marc.info/?l=netfilter-devel&m=162038133102042&w=2

Also, given the first element is always the same (the virtual CIDR) we don't
really need to concatenation. We can rewrite the ruleset in a more compact
fashion, with a few early checks for additional housekeeping.

Bug: T270704
Change-Id: Ib1e0429ca0a29832a9f84a4a4b83b224b41c2e6f
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>