toolforge: re-enable toolforge certificate monitor
It appears we hastily removed our monitor when we switched to acme-chief.
Recent TLS issues suggest the monitor is still a very good idea.

I've added the admin tool because I'm not sure if a redirect will work.
If it does work with our validator, then I'll just use the main domain.
We could use something for wmcloud.org as well, but what?

Bug: T282264
Change-Id: Ibffdff9805e379a9345bb6ecb1461d38435f77bd