peopleweb: put a public_html into /etc/skel to ensure all users get one
The docs at https://wikitech.wikimedia.org/wiki/People.wikimedia.org#How_to_upload_files
currently have a step:

"Put files into a directory called "public_html".
If that doesn't exist yet, create it with mkdir."

This is to remove that second part and automatically create one
for each user, but of course only on hosts using the peopleweb role.

Simply creating the directory in /etc/skel should do it, as long
as useradd runs with -m and creates the home dirs.

It should be root owned and puppet should automatically handle
permissions for a directory so defaults should be fine.

Of course this will only do something on a new host and not
retroactively on already running hosts and I don't want to reimage
just for this though.

Bug: T280989
Change-Id: Iacdcd978de9417329ce7153900d9e4048ef7d073