admin: add jhuneidi to contint-roots
Add Jeena to contint-roots.

contint-roots was originally created for contint* servers,
so it gives root access there.

But later it has also been applied to releases* servers for
the reasons that is also given in the request:

Reason for access: Needed for docker image building work for mw-on-k8s

Already has approval from Tyler, who is also the "approval: " person
for this admin group specifically.

And of course it's obvious that Jeena is involved heavily in contint things.

Since she is an existing shell user with memberships including releasers-mwcli,
releasers-blubber and deployment this should as easy as adding to the group
(and maybe running puppet on releases*).

Bug: T282610
Change-Id: Iac0ddaf8d6be0e67eb4e7378da92a6e52848d9ec