docker-registry: Clean up old nginx http endpoint
Why:
We have an nginx powered docker-registry endpoint that is currently
unused. Starting with b788e69b652 we have initiated a process to move to
the HTTPS only counterpart. The HTTP only counterpart was kept around
just for varnish as it wasn't able to use HTTPS. By now, the connections
to the services are being done by ats-backend and that's HTTPS compliant
so we can finally get rid of the HTTP only nginx configuration

What:
* Remove the http_endpoint puppet parameter
* Unconditionally absent the configuration
* Remove the file

Bug: T256762
Change-Id: I08cdcf392da14f07a5f3b352ddcad2ec0002a6c6