docker-registry: Re-apply Cache-Control rules
Why:
In 3a804db23d3, we removed the varnish only HTTP registry endpoint.
Migrate the 1 important rule of it in the HTTPS endpoints. This is
expected to solve a variety of issues with cacheability that are
documented in T256762.

What:
Make sure we don't cache anything by setting no-cache,must-revalidate in
cache-control. But DO make an exception for the catalog, allowing it to
be cached for up to 5minutes. The catalog is expensive to calculate and
rather small in size as a response so that should be fine.

The internal usages of the registry should not be affected by this. The
restricted area is also on purpose not being changed as it's only mean
to be internally accessible

Bug: T256762
Change-Id: I556ca96dc2cb567d619c5614b5fa19d68ff095a8