cloud nfs: Change primary cluster rate limits dramatically
In the past (specifically for I6142677e26fe41f3599968357), the read limit
was dropped extremely low without explanation. Presumably, this was to
free bandwidth for writes, which were having serious issues at the time.

The values were never adjusted back up when the cluster changed over time.
At this point, the NFS server and all clients are on 10Gb Ethernet.

There is no way the existing values were sane as a result. Bumping both
up to (naturally) 10x the value for what writes would be.

Bug: T218338
Change-Id: I5b59cf2438c42b37fd0d2e600a1fe56b783edc7c