varnish: fix check_vcl_reload state check
Although the check_vcl_reload script should have failed on the whole
eqsin cluster due to a misalignment between the list of nodes available
in etcd and those in hiera, the check keeps saying OK. This is due to ==
being used without spaces around it in the conditional looking for the
contents of /var/run/reload-vcl-state. Also, == is a bashism and =
should be used instead.

Bug: T282880
Change-Id: I7d40fb7ff23bc43f7e4fc55e57ac1b4b0bc76b7d