wmfkeystonehooks: Narrow our security group search to the new project
Previously it was enough to have project-specific credentials
when calling list_security_groups.  Now that novaadmin has global
admin, though, that call now defaults to showing all security groups
for all projects.

Limiting this search to the actual newly created project allows us to
still pick the first 'default' security group and have it be the one
we actually want to edit.

Bug: T282876
Change-Id: Ie1a9babe0a96d6ba2e260f28b5bac3022763b72c