install_server: add buster-rescue tftp config
currently rescue is an option in the network debian installer menu,
but if no selection is made before the timeout the server will be reimaged.

this adds a rescue only netboot config, for instances where we want to be
sure the host will pxe boot into rescue mode, without risking a reimage.

also, attempt to resurrect mwlog1001 using buster-rescue in order
to pull files from /home.

Bug: T282575
Change-Id: I2506e667caa5e8a8ff465fcad6c2abfd1f6d3783