registry: Add proxy_pass to the catalog location block
Why:
In I556ca96dc2cb567d619c5614b5fa19d68ff095a8 we reapplied the
Cache-Control directives and even added an optimization for allowing to
cache /v2/_catalog as it is a rather expensive endpoint. This did not
work and would return 404.

This is confusing but it appears that proxy_pass is what is called an
"action directive". I could find no list of those directives, they are
defined as the directives that perform some action (rewrite is the usual
example) and not just set some configuration. Those are reportedly not
inherited between nested location blocks and need to be redefined so
doing so in this commit. It's unclear if proxy_set_header are action
directives, that will need some more testing.

What:
Move the catalog matching direct at the end for more clarity and add the
proxy_pass directive. That should fix the regression

Bug: T256762
Change-Id: I690a87dbfca8485881aad94cceb734864b3f318d