acme_chief: add certificates for wikimedia-dns.org
The new domain for Wikidough will be wikimedia-dns.org so we generate
new certificates for that from the doh200* hosts.

To maintain backward compatibility with malmok.wikimedia.org before we
switch all users to the new domain, we add it to the SAN for the new
certificates. The malmok host (and the domain) will be decommissioned
later in favour of the new domain and doh* hosts.

Bug: T252132
Change-Id: Id527054bbe387dc94c3b9b0387caa9489a8f81f5
Hosts: O:acme_chief