Allow absenting the helmchartctl systemd timer
Why:
On T283294 we formulated the hypothesis that this is due to a race
condition between the 2 systemd timers on eqiad and codfw chartmuseum
hosts running simultaneously. While this remains our best hypothesis,
and we should aim to try and resolve it correctly at the level of the
race condition, probably via some timer mangling, let's first validate
the hypothesis that it's indeed the fact that those 2 job running
concurrently is what end's up corrupting the packaged charts.

What:
Allow disabling the systemd timer selectively. This is so temporary as a
stopgap (it's really just to validate the hypothesis), that we just set
the hiera flag on the 2 hosts

Once we have verification (we should just bump a couple of chart
versions and test soon after this is merged), the entirety of this
commit should be reverted

Bug: T283147
Change-Id: If6d3dca1530af70e0b85854c920a2c4c32a7fd3a